Refactored jobs CRUD with connection class and 4 methods
In this commit, I refactored the jobs CRUD functionality by creating a connection class and implementing 4 methods. These methods include one that retrieves all elements from the table, one that deletes an element by ID, one that updates an element, and one that selects an element by ID. This refactoring should improve the maintainability and organization of the code.