Employees crud finished
Finished implementing employee CRUD operations for the application. All CRUD functionality has been thoroughly tested and verified to be working as expected.

In addition, the database schema was updated to use MEDIUMBLOB fields for storing larger files and images, allowing for greater flexibility and scalability when working with larger data sets. This update required modifying existing tables and columns, as well as updating the application code to work with the new schema.